converted linalg to a submodule, split qr into indv file+refactor names, created svd file
