+++ heat/core/linalg/__init__.py
@@ -1,0 +1,2 @@
+from . import basics
+from . import qr
